Handle hidden methods
